Major: Function object to output plane data
The function object cuttingPlane outputs .csv files
that are plot as a countour map by plot_surf.py.

The file yplane is an example file, that should be included
in the system controlDict.

The csv files are stored in the postProcessing dir that
exists in the case dir for serial runs and the processor
dirs for parallel runs.